Rename config folder to hydrazen
