--- src/configs/__init__.py
@@ -1,3 +1,0 @@
-import src.configs.experiment as experiment
-
-__all__ = ["experiment"]
