--- src/configs/experiment/__init__.py
@@ -1,3 +1,0 @@
-import src.configs.experiment.mnist as mnist
-
-__all__ = ["mnist"]
