+++ src/hydrazen/__init__.py
@@ -1,0 +1,3 @@
+import src.hydrazen.experiment as experiment
+
+__all__ = ["experiment"]
