+++ src/hydrazen/experiment/__init__.py
@@ -1,0 +1,3 @@
+import src.hydrazen.experiment.mnist as mnist
+
+__all__ = ["mnist"]
